pyproject.toml
[build-system]
requires = ["pdm-pep517>=1.0"]
build-backend = "pdm.pep517.api"

[project]
name = "mit-ub-chexpert"
description = ""
requires-python = ">=3.11,<3.12"
authors = [
    {name = "Scott Chase Waggener", email = "tidalpaladin@protonmail.com"},
]
dependencies = [
  "numpy",
  "Pillow",
  "pandas",
  "torch>=2.4",
  "dicom-preprocessing @ git+https://github.com/medcognetics/dicom-preprocessing.git",
  "mit-ub @ git+https://github.com/TidalPaladin/mit-ub.git@feat/conv-transformer",
]


readme = "README.md"
license = {text = "Apache"}
dynamic = ["version"]

[project.optional-dependencies]
train = [
    "mit-ub[train] @ git+https://github.com/TidalPaladin/mit-ub.git",
    "Pillow",
    "pandas",
    "pytorch-lightning>=2.4",
    "torchvision",
    "torchmetrics",
    "bitsandbytes>=0.44.1",
    "jsonargparse[signatures]",
    "omegaconf",
    "tqdm-multiprocessing",
    "jsonargparse[signatures]",
    "deep-helpers @ git+https://github.com/TidalPaladin/deep-helpers.git",
    "torch-dicom[datamodule] @ git+https://github.com/medcognetics/torch-dicom.git",
]

[tool.autoflake]
remove-all-unused-imports = true
remove-unused-variables = true

[tool.autopep8]
max_line_length = 120
ignore = "E501,W6,E203"
in-place = true
recursive = true
aggressive = 3

[tool.black]
line-length = 120

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 119
lines_after_imports = 2
multi_line_output = 3
use_parentheses = false

[tool.pyright]
include = ["project", "tests"]
exclude = [
  "**/node_modules",
  "**/__pycache__",
  "**/.pytest_cache",
]
ignore = ["src/oldstuff"]
pythonVersion = "3.11"
pythonPlatform = "Linux"

[tool.pdm]
version = { source = "scm" }

[tool.pdm.dev-dependencies]
test = [
  "pytest",
  "pytest-mock",
  "pytest-cov",
  "coverage",
  "pdbpp",
]
quality = [
  "autoflake",
  "autopep8",
  "black",
  "flake8",
  "isort",
]

[tool.pytest.ini_options]
markers = [
  "ci_skip",
]
filterwarnings = []

# Sample script config
#[project.scripts]
#foo = "project.cli.__main__:main"
